fix(secure-storage): only trim by the aws secret manager global prefix to get the resulting key
This follows up on
https://github.com/bf2fc6cc711aee1a0c2a/kas-fleet-manager/pull/1638/commits/64d56dd8ccb7899eca2c888f5e7b01286a89d702

which introduce an unintended bug of returning wrong keys as expected by certmagic since prefix was removed.

For example:

If the storage.secretPrefix is "kas-fleet-manager"

When we call `Store(ctx, "example/secret-one")` then the resulting secret name in the storage backend is going to be "kas-fleet-manager/example/secret-one".
When we call `List(ctx, "example", true)` then the returned list of keys should be `["example/secret-one"]` and not `/secret-one` as it is current is.
This was because we trimmed by the filterValue, an exising bug from the first version. We should only trim by the "kas-fleet-manager/" prefix.