chore: remove observatorium auth type flag
Since the removal of Observatorium DEX authentication only OIDC
is supported. Because of this we remove the auth type flag
to simplify the configuration around Observatorium. If in
the future we start supporting additional authentication methods
for Observatorium we can reintroduce the functionality but
possibly redesigned in a different/better way.