feat: add markdown link checking tool and related make targets
This set of changes adds two Makefile targets
* markdown-link-check-install: installs the markdown-link-check
  tool which is able to check whether links are alive
* lint/markdown-link-check: Runs markdown-link-check checking
  all .md (markdown) files in the project. The bin directory is
  excluded

markdown-link-check is able to check:
* URLs
* Relative links to other files in the projects
* `mailto:` links

Additionally, a markdown-link-check configuration file
`.markdown-link-check/markdown-link-check-config.json` relative
to the project's root has been added. This file can be used
when running the markdown-link-check tool with the `-c` flag
referencing it. The `lint/markdown-link-check` Makefile target
makes use of it automatically.

Some URL patterns have been configured to be ignored in the
markdown-link-check configuration file. This URL patterns
correspond to URLs that resolve to private IP addresses or that
are private GitHub/GitLab repositories.